% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/whoami.R
\name{email_address}
\alias{email_address}
\title{Email address of the current user}
\usage{
email_address()
}
\value{
Email address on success. Otherwise an error is thrown.
}
\description{
It tries to find it in the user's global git configuration.
}
\examples{
\dontrun{
email_address()
}
}
\seealso{
Other user.names: \code{\link{fullname}};
  \code{\link{gh_username}}; \code{\link{username}};
  \code{\link{whoami}}
}

