% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/whoami.R
\name{whoami}
\alias{whoami}
\title{User name and full name of the current user}
\usage{
whoami()
}
\value{
A named character vector with two entries: \code{username} and
  \code{fullname}.
}
\description{
Calls \code{\link{username}} and \code{\link{fullname}}.
}
\examples{
whoami()
}
\seealso{
Other user.names: \code{\link{fullname}};
  \code{\link{username}}
}

