% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whoami.R
\name{email_address}
\alias{email_address}
\title{Email address of the current user}
\usage{
email_address(fallback = NULL)
}
\arguments{
\item{fallback}{If not \code{NULL} then this value is returned
if the email address cannot be found, instead of triggering an error.}
}
\value{
Email address on success. Otherwise an error is thrown.
}
\description{
If uses the \code{EMAIL} environment variable, if set.
Otherwise it tries to find it in the user's global git configuration.
}
\examples{
\dontrun{
email_address()
}
}
\seealso{
Other user names: \code{\link{fullname}},
  \code{\link{gh_username}}, \code{\link{username}},
  \code{\link{whoami}}
}
\concept{user names}
