% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whoami.R
\name{fullname}
\alias{fullname}
\title{Full name of the current user}
\usage{
fullname(fallback = NULL)
}
\arguments{
\item{fallback}{If not \code{NULL} then this value is returned
if the full name of the user cannot be found, instead of
triggering an error.}
}
\value{
The full name of the current user.
}
\description{
Uses the \code{FULLNAME} environment variable, if set.
Otherwise tries system full names and the git configuration as well.
}
\examples{
\dontrun{
fullname()
}
}
\seealso{
Other user names: \code{\link{email_address}},
  \code{\link{gh_username}}, \code{\link{username}},
  \code{\link{whoami}}
}
\concept{user names}
