% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/whoami.R
\name{username}
\alias{username}
\title{User name of the current user}
\usage{
username()
}
\value{
The user name of the current user.
}
\description{
Tries the \code{LOGNAME}, \code{USER}, \code{LNAME}, \code{USERNAME}
environment variables first. Then it tries the `id` command on Unix-like
platforms and `whoami` on Windows.
}
\examples{
\dontrun{
username()
}
}
\seealso{
Other user.names: \code{\link{email_address}};
  \code{\link{fullname}}; \code{\link{gh_username}};
  \code{\link{whoami}}
}

