% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whoami-package.R
\docType{package}
\name{whoami-package}
\alias{whoami}
\alias{whoami-package}
\title{whoami: Username, Full Name, Email Address, 'GitHub' Username of the Current User}
\description{
Look up the username and full name of the current user,
    the current user's email address and 'GitHub' username,
    using various sources of system and configuration information.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/whoami#readme}
  \item Report bugs at \url{https://github.com/r-lib/whoami/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Other contributors:
\itemize{
  \item Maëlle Salmon [contributor]
}

}
\keyword{internal}
