% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whoami.R
\name{whoami}
\alias{whoami}
\title{User name and full name of the current user}
\usage{
whoami()
}
\value{
A named character vector with entries: \code{username},
  \code{fullname}, \code{email_address}, \code{gh_username}.
}
\description{
Calls \code{\link{username}} and \code{\link{fullname}}.
}
\examples{
\dontrun{
whoami()
}
}
\seealso{
Other user names: \code{\link{email_address}},
  \code{\link{fullname}}, \code{\link{gh_username}},
  \code{\link{username}}
}
